Windows: Initial compilation enablement
