+++ cgroups/fs/cpuset_test.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package fs
 
 import (
