+++ cgroups/fs/freezer.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package fs
 
 import (
