+++ cgroups/fs/fs_unsupported.go
@@ -1,0 +1,3 @@
+// +build !linux
+
+package fs
