+++ cgroups/fs/stats_util_test.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package fs
 
 import (
