+++ cgroups/fs/util_test.go
@@ -1,3 +1,5 @@
+// +build linux
+
 /*
 Utility for testing cgroup operations.
