+++ configs/device_defaults.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package configs
 
 var (
