+++ stats.go
@@ -1,12 +1,5 @@
 package libcontainer
 
-import "github.com/docker/libcontainer/cgroups"
-
-type Stats struct {
-	Interfaces  []*NetworkInterface
-	CgroupStats *cgroups.Stats
-}
-
 type NetworkInterface struct {
 	// Name is the name of the network interface.
 	Name string
