+++ stats_linux.go
@@ -1,0 +1,8 @@
+package libcontainer
+
+import "github.com/docker/libcontainer/cgroups"
+
+type Stats struct {
+	Interfaces  []*NetworkInterface
+	CgroupStats *cgroups.Stats
+}
