+++ system/sysconfig_notcgo.go
@@ -1,8 +1,15 @@
-// +build !cgo
+// +build !cgo windows
 
 package system
 
 func GetClockTicks() int {
 	// TODO figure out a better alternative for platforms where we're missing cgo
+	//
+	// TODO Windows. This could be implemented using Win32 QueryPerformanceFrequency().
+	// https://msdn.microsoft.com/en-us/library/windows/desktop/ms644905(v=vs.85).aspx
+	//
+	// An example of its usage can be found here.
+	// https://msdn.microsoft.com/en-us/library/windows/desktop/dn553408(v=vs.85).aspx
+
 	return 100
 }
